Fix Grouped Instances counter
Enhances flavor detail management, including callbacks and helper
process. This Fixes #25.
